jira-confluence-mcp-main/pyproject.toml
[project]
name = "confluence-mcp"
version = "0.1.0"
description = "Confluence MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atlassian-python-api>=4.0.4",
    "mcp>=1.9.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]
